requirements.txt
django
django-mptt
Faker